+++ app/src/org/commcare/activities/StandardHomeActivityUIController.java
@@ -9,6 +9,7 @@ import android.widget.TextView;
 import androidx.annotation.ColorRes;
 import androidx.annotation.StringRes;
 import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.widget.AppCompatButton;
 import androidx.cardview.widget.CardView;
 import androidx.core.content.ContextCompat;
 import androidx.recyclerview.widget.LinearLayoutManager;
@@ -22,14 +23,17 @@ import org.commcare.adapters.HomeScreenAdapter;
 import org.commcare.android.database.connect.models.ConnectAppRecord;
 import org.commcare.android.database.connect.models.ConnectDeliveryPaymentSummaryInfo;
 import org.commcare.android.database.connect.models.ConnectJobRecord;
+import org.commcare.android.database.connect.models.ConnectTaskRecord;
 import org.commcare.connect.ConnectDateUtils;
 import org.commcare.connect.ConnectJobHelper;
+import org.commcare.connect.ConnectNavHelper;
 import org.commcare.connect.database.ConnectJobUtils;
+import org.commcare.connect.database.ConnectTaskUtils;
 import org.commcare.dalvik.R;
 import org.commcare.interfaces.CommCareActivityUIController;
+import org.commcare.personalId.UnlockPolicy;
 import org.commcare.preferences.DeveloperPreferences;
 import org.commcare.preferences.HiddenPreferences;
-import org.commcare.connect.database.ConnectTaskUtils;
 import org.commcare.suite.model.Profile;
 
 import java.util.ArrayList;
@@ -49,6 +53,7 @@ public class StandardHomeActivityUIController implements CommCareActivityUIContr
     private View viewJobCard;
     private CardView connectMessageCard;
     private ImageView connectMessageWarningIcon;
+    private AppCompatButton acbOpenConversation;
     private ConnectProgressJobSummaryAdapter connectProgressJobSummaryAdapter;
 
     private HomeScreenAdapter adapter;
@@ -85,6 +90,7 @@ public class StandardHomeActivityUIController implements CommCareActivityUIContr
         viewJobCard = activity.findViewById(R.id.viewJobCard);
         connectMessageCard = activity.findViewById(R.id.cvConnectMessage);
         connectMessageWarningIcon = activity.findViewById(R.id.ivConnectMessageWarningIcon);
+        acbOpenConversation = viewJobCard.findViewById(R.id.acb_open_conversation);
         connectProgressJobSummaryAdapter = new ConnectProgressJobSummaryAdapter(new ArrayList<>());
         RecyclerView recyclerView = viewJobCard.findViewById(R.id.rdDeliveryTypeList);
         recyclerView.setLayoutManager(new LinearLayoutManager(activity));
@@ -134,6 +140,25 @@ public class StandardHomeActivityUIController implements CommCareActivityUIContr
             cvRelearnTasksPending.setVisibility(View.GONE);
             tvJobTime.setVisibility(View.GONE);
             hoursTitle.setVisibility(View.GONE);
+        }
+        setOpenConversationButtonVisibility(job);
+    }
+
+    private void setOpenConversationButtonVisibility(ConnectJobRecord job) {
+        ConnectTaskRecord messagingTask = ConnectTaskUtils.getValidPendingOcsTask(activity, job);
+        if (messagingTask == null) {
+            acbOpenConversation.setVisibility(View.GONE);
+        } else {
+            acbOpenConversation.setVisibility(View.VISIBLE);
+            acbOpenConversation.setOnClickListener(v ->
+                    ConnectNavHelper.INSTANCE.unlockAndGoToMessaging(
+                            activity,
+                            UnlockPolicy.SESSION_WITH_TIME_THRESHOLD,
+                            messagingTask.getConnectChannelId(),
+                            (success, error) -> {
+                            }
+                    )
+            );
         }
     }