+++ app/src/org/commcare/connect/database/ConnectTaskUtils.kt
@@ -8,6 +8,7 @@ import org.commcare.models.database.SqlStorage
 import org.commcare.preferences.ConnectJobPreferences
 import org.commcare.utils.SyncDetailCalculations
 import org.javarosa.core.model.utils.DateUtils
+import org.javarosa.core.services.Logger
 import java.util.Date
 
 /**
@@ -50,7 +51,7 @@ object ConnectTaskUtils {
         }
 
         ConnectJobUtils.getJobPreferences(jobUUID).apply {
-            setRelearnTaskPending(incoming.any { it.status == "assigned" })
+            setRelearnTaskPending(incoming.any { it.status == ConnectTaskRecord.STATUS_ASSIGNED })
             resetRelearnTasksCompletedTime()
             if (changed) {
                 updateTaskModifiedTime()
@@ -69,7 +70,7 @@ object ConnectTaskUtils {
     ): Boolean {
         val tasks = getTasksForJob(context, jobUUID, null)
         if (tasks.isNotEmpty()) {
-            return tasks.any { it.status == "assigned" }
+            return tasks.any { it.status == ConnectTaskRecord.STATUS_ASSIGNED }
         }
         return ConnectJobUtils.getJobPreferences(jobUUID).isRelearnTaskPending()
     }
@@ -86,7 +87,25 @@ object ConnectTaskUtils {
         context: Context,
         jobUUID: String,
         type: String,
-    ): ConnectTaskRecord? = getTasksForJob(context, jobUUID, null).find { it.type == type && it.status == "assigned" }
+    ): ConnectTaskRecord? =
+        getTasksForJob(context, jobUUID, null).find { it.type == type && it.status == ConnectTaskRecord.STATUS_ASSIGNED }
+
+    @JvmStatic
+    fun getValidPendingOcsTask(
+        context: Context,
+        job: ConnectJobRecord,
+    ): ConnectTaskRecord? {
+        if (job.status != ConnectJobRecord.STATUS_DELIVERING) return null
+        val task = getPendingTaskOfType(context, job.jobUUID, ConnectTaskRecord.TYPE_OCS) ?: return null
+        if (task.connectChannelId.isEmpty()) {
+            Logger.exception(
+                "Invalid messaging task",
+                Throwable("Messaging task has no channel id: ${task.taskId}"),
+            )
+            return null
+        }
+        return task
+    }
 
     @JvmStatic
     fun shouldShowTasksCompletedMessage(
@@ -119,7 +138,7 @@ object ConnectTaskUtils {
     ): ConnectTaskRecord? {
         val tasks = getTasksForJob(context, jobUUID, null)
         if (tasks.isNotEmpty()) {
-            return tasks.filter { it.status == "completed" }.maxByOrNull { it.dateModified }
+            return tasks.filter { it.status == ConnectTaskRecord.STATUS_COMPLETED }.maxByOrNull { it.dateModified }
         }
         val prefs = ConnectJobUtils.getJobPreferences(jobUUID)
         val completedTimeMs = prefs.getRelearnTasksCompletedTimeMs()